Enforce saving of input image in mgz format in eval.py
For uniformity, the format of the original input image is currently
restricted to be in mgz (subject to change).
Also, if the input image was not originally a mgz (for e.g. nii), the
image saving would fail if a MGHImage object is not explictly created.
This logic may change after incorporating the new save_image
functionality.