[run_prediction] Remove faulty vox_size str and numeric checks
The isnumeric() and str checks will always fail for a float input
(decimal number), e.g. 0.8.
For the str condition, the definition of the vox_size function in
arg_types.py means that it can be a float and not necessarily always a string.
Therefore, having isinstance(x, str) leads to any float value being flagged as
an invalid value.