[run_prediction] Set dtype to uchar when calling save_image
This fixes certain failure cases with nifti images in which the
conformed always gets saved with a wrong dtype (int32), despite
both the image array and header being explicitly set to uchar.
This leads to the same image failing the conform check at the beginning
of the surf pipeline.

The fix is to ensure that set_data_dtype() is used on the MGHImage object
itself, which is now done in the imported data_utils function.