The version script had a hard-coded flag to never use git.
This forced the build_script to not use git in its build script to generate a new BUILD.info file.
Which in turn meant that the build script tried to read from the BUILD.info file to create a new BUILD.info file.
This lead to a circular dependency and a crash during build.

This Commit fixes this issue by:
- removing the hard-coded git flag
- adding an option to the version script to NOT read the BUILD.info file