Fix docker build script, only works with docker-container builder, if buildx is installed.
Fix checks for the buildx/docker-container requirements. Used to always require docker-container builder, if buildx was installed. This requirement is a bit too strict. The docker builder is good enough, if no attestation or caching are selected.
Fix image exporter type, if building with --attest (attestation is not supported by the docker exporter).
Restructure code to make it readable, understandable and maintainable.