Edit and establish the workflow "quicktest" on github actions servers (compatible with github actions)
The quicktest action is triggered manually (dispatch in actions) or directly from a PR. The former can for example be used to generate the reference data. The latter may be used to add on-demand checks.

The actual test scripts have not been modified and are not expected to seamlessly work, but processing works.

On-demand checks in the pull request are triggered by adding or removing a "quicktest" label. There is a further check to ensure the workflow is only performed if people listed as collaborators changed the label.

This Commit adds/changes the github workflow files and adds actions where appropriate.
Github actions are organized in .github/actions
- build-docker
- load-docker
- load-processed
- run-fastsurfer
- run-tests

The core workflow is .github/workflows/quicktest.yml. It requires 5 secrets to be set up in the repository which include the FreeSurfer license and paths to two sample images as well as their reference processing.
The Workflow defines several jobs to enable parallel processing as well as circumvent resource limitations of github actions.