Fixes additional flags in srun_fastsurfer
Additional flags (including --t2) did not work properly from the subject_list file in srun_fastsurfer.sh
The Bug is actually in stools.sh:translate_cases, where the lines are not correct parsed.
This Commit fixes this issue and properly parses the file.
Also, the check to find non-existent files in stools.sh:check_subject_images.

This Commit also adds a check if files are symlinks, which may cause problems with folders mounted in singularity (this needed a small change in srun_fastsurfer.sh - an additional argument to check_subject_images).